grammar: move into project_loader
This is the first step toward making the grammar more generic and used
in more of the YAML. Other than moving the grammar around (and the small
import tweaks necessary to continue working) this commit makes no
changes.

Updates: #1492
