+++ setup.py
@@ -35,8 +35,8 @@ packages = [
 	'snapcraft.internal.cache',
 	'snapcraft.internal.deltas',
 	'snapcraft.internal.pluginhandler',
-	'snapcraft.internal.pluginhandler.stage_package_grammar',
 	'snapcraft.internal.project_loader',
+	'snapcraft.internal.project_loader.grammar',
 	'snapcraft.internal.repo',
 	'snapcraft.internal.sources',
 	'snapcraft.internal.states',
