+++ snapcraft/internal/pluginhandler/_stage_package_handler.py
@@ -17,7 +17,7 @@
 import logging
 
 from snapcraft.internal import repo
-from snapcraft.internal.pluginhandler import stage_package_grammar
+from snapcraft.internal.project_loader import grammar
 
 logger = logging.getLogger(__name__)
 
@@ -69,7 +69,7 @@ class StagePackageHandler:
     def _stage_packages(self):
         # Comparing to None here since after calculation it may be an empty set
         if self.__stage_packages is None:
-            self.__stage_packages = stage_package_grammar.process_grammar(
+            self.__stage_packages = grammar.process_grammar(
                 self._grammar, self._project_options, self._repo)
 
         return self.__stage_packages
